langGraph/longterm/pyproject.toml
[project]
name = "longterm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.8",
    "openai>=2.2.0",
    "python-dotenv>=1.1.1",
]
